backend/pyproject.toml
[project]
name = "quart_project"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "dana367", email = "dana.larina1@gmail.com"},
]
dependencies = [
    "quart>=0.19.6",
    "pytest-asyncio>=0.24.0",
    "quart-db>=0.8.3",
    "quart-auth>=0.10.1",
    "bcrypt>=4.2.0",
    "zxcvbn>=4.4.28",
    "itsdangerous>=2.2.0",
    "pydantic[email]>=2.9.2",
    "quart-rate-limiter>=0.10.0",
    "quart-schema>=0.20.0",
    "httpx>=0.27.2",
    "flake8-print>=5.0.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "Private"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true



[tool.pdm.dev-dependencies]
dev = [
    "black>=24.8.0",
    "isort>=5.13.2",
    "djhtml>=3.0.6",
    "mypy>=1.11.2",
    "flake8>=7.1.1",
    "bandit>=1.7.9",
    "vulture>=2.11",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]

[[tool.mypy.overrides]]
module =["h11"]
ignore_missing_imports = true

[tool.pdm.scripts]
start = {cmd = "quart --app src/backend/run.py run --port 5050", env_file = "development.env"}
format-black = "black src/ tests/"
format-djhtml = "djhtml src/backend/templates -t 2 --in-place"
format-isort = "isort src tests --apply"
format = {composite = ["format-black", "format-djhtml", "format-isort"]}
lint-bandit = "bandit -r src/"
lint-black = "black --check --diff src/ tests/"
lint-djhtml = "djhtml src/backend/templates -t 2 --check"
lint-flake8 = "flake8 src/ tests/"
lint-isort = "isort --check --diff src tests"
lint-mypy = "mypy src/backend tests"
lint-vulture = "vulture src"
lint = {composite = ["lint-flake8", "lint-bandit", "lint-black", "lint-djhtml", "lint-isort", "lint-mypy", "lint-vulture"]}
recreate-db-base = "quart --app src/backend/run.py recreate_db"
recreate-db = {composite = ["recreate-db-base"], env_file = "development.env"}
test = {composite = ["recreate-db-base", "pytest tests/"], env_file = "testing.env"}

[tool.black]
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.vulture]
min_confidence = 100

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"  # You can change the scope to session, module, etc.
addopts = "--showlocals"
asyncio_mode = "auto"
pythonpath = ["src"]